requirements.txt
praw==7.7.1
python-dotenv==1.0.0
openai>=1.1.0
pydantic>=2.0.0
tqdm>=4.65.0 